requirements.txt
numpy
pyserial-asyncio-fast
aiohttp
aiortc
shapely